requirements.txt
pymongo
python-dotenv
python-telegram-bot
google-generativeai
RestrictedPython